+++ eslint/es6/index.js
@@ -3,6 +3,9 @@ import security from 'eslint-plugin-security';
 import jsdoc from 'eslint-plugin-jsdoc';
 import stylistic from '@stylistic/eslint-plugin';
 import importPlugin from 'eslint-plugin-import';
+import promise from 'eslint-plugin-promise';
+import sonarjs from 'eslint-plugin-sonarjs';
+import unicorn from 'eslint-plugin-unicorn';
 import rules from './rules.js';
 
 export default [
@@ -18,7 +21,10 @@ export default [
         plugins: {
             security,
             jsdoc,
-            '@stylistic': stylistic
+            '@stylistic': stylistic,
+            promise,
+            sonarjs,
+            unicorn
         },
         rules
     }